+++ tests/common/features/schema.rs
@@ -43,6 +43,7 @@ pub async fn create_tables(db: &DatabaseConnection) -> Result<(), DbErr> {
     create_insert_default_table(db).await?;
     create_pi_table(db).await?;
     create_uuid_fmt_table(db).await?;
+    create_edit_log_table(db).await?;
 
     if DbBackend::Postgres == db_backend {
         create_collection_table(db).await?;
@@ -335,6 +336,12 @@ pub async fn create_json_struct_table(db: &DbConn) -> Result<ExecResult, DbErr> 
 }
 
 pub async fn create_collection_table(db: &DbConn) -> Result<ExecResult, DbErr> {
+    db.execute(sea_orm::Statement::from_string(
+        db.get_database_backend(),
+        "CREATE EXTENSION IF NOT EXISTS citext".into(),
+    ))
+    .await?;
+
     let stmt = sea_query::Table::create()
         .table(collection::Entity)
         .col(
@@ -343,6 +350,11 @@ pub async fn create_collection_table(db: &DbConn) -> Result<ExecResult, DbErr> {
                 .not_null()
                 .auto_increment()
                 .primary_key(),
+        )
+        .col(
+            ColumnDef::new(collection::Column::Name)
+                .custom(Alias::new("citext"))
+                .not_null(),
         )
         .col(
             ColumnDef::new(collection::Column::Integers)
@@ -468,4 +480,21 @@ pub async fn create_uuid_fmt_table(db: &DbConn) -> Result<ExecResult, DbErr> {
         .to_owned();
 
     create_table(db, &stmt, UuidFmt).await
+}
+
+pub async fn create_edit_log_table(db: &DbConn) -> Result<ExecResult, DbErr> {
+    let stmt = sea_query::Table::create()
+        .table(edit_log::Entity)
+        .col(
+            ColumnDef::new(edit_log::Column::Id)
+                .integer()
+                .not_null()
+                .auto_increment()
+                .primary_key(),
+        )
+        .col(ColumnDef::new(edit_log::Column::Action).string().not_null())
+        .col(ColumnDef::new(edit_log::Column::Values).json().not_null())
+        .to_owned();
+
+    create_table(db, &stmt, EditLog).await
 }